OperatorService Uptdate to generic service

+++ back-end/src/main/java/rccommerce/controllers/OperatorController.java
@@ -14,7 +14,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
@@ -22,6 +21,7 @@ import jakarta.validation.Valid;
 import rccommerce.dto.OperatorDTO;
 import rccommerce.dto.OperatorMinDTO;
 import rccommerce.services.OperatorService;
+import rccommerce.services.exceptions.ForbiddenException;
 
 @RestController
 @RequestMapping(value =  "/operators")
@@ -31,13 +31,21 @@ public class OperatorController {
 	private OperatorService service;
 	
 	@PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_OPERATOR', 'ROLE_SELLER')")
-	@GetMapping
+	@GetMapping(value = "/all")
 	public ResponseEntity<Page<OperatorMinDTO>> findAll(
-			@RequestParam(name = "name", defaultValue = "") String name,
-			@RequestParam(name = "email", defaultValue = "") String email,
+//			@RequestParam(name = "name", defaultValue = "") String name,
+//			@RequestParam(name = "email", defaultValue = "") String email,
 			Pageable pageable) {
-		Page<OperatorMinDTO> dto = service.findAll(name, email, pageable);
+		Page<OperatorMinDTO> dto = service.findAll(pageable);
 		return ResponseEntity.ok(dto);
+	}
+	
+	@PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_OPERATOR', 'ROLE_SELLER', 'ROLE_CLIENT')")
+	@GetMapping(value = "/search")
+	public ResponseEntity<Page<OperatorMinDTO>> seachOperator(
+			@Valid @RequestBody OperatorMinDTO dto, Pageable pageable) {
+		Page<OperatorMinDTO> pageDto = service.searchOperator(dto, pageable);
+		return ResponseEntity.ok(pageDto);
 	}
 	
 	@PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_OPERATOR', 'ROLE_SELLER')")
@@ -66,6 +74,9 @@ public class OperatorController {
 	@PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_OPERATOR', 'ROLE_SELLER')")
 	@DeleteMapping(value = "/{id}")
 	public ResponseEntity<Void> delete(@PathVariable Long id) {
+		if (id == 1L) {
+			throw new ForbiddenException("ADMINSTRADOR MASTER - Deleção proibida");
+		}
 		service.delete(id);
 		return ResponseEntity.noContent().build();
 	}
